requirements.txt
requests
discord.py
soundcloud-lib
youtube-dl